refactor: move classes to new packages
