+++ src/main/java/edu/ntnu/idi/idatt/domain/entity/Dice.java
@@ -1,4 +1,4 @@
-package edu.ntnu.idi.idatt.model.entities;
+package edu.ntnu.idi.idatt.domain.entity;
 
 import edu.ntnu.idi.idatt.utils.Validation;
